+++ apps/desktop/src/main/__tests__/renderer-style-pruning-contract.test.ts
@@ -12,12 +12,14 @@ describe('renderer style pruning contract', () => {
     const retiredHooks = [
       'maka-indeterminate-bar',
       'maka-nav-disclosure',
+      'maka-nav-primary',
       'maka-nav-tree',
       'maka-session-archive-link',
       'maka-session-filter',
       'maka-session-panel-help-chip',
       'maka-session-search-clear',
       'maka-sidebar-brand',
+      'maka-skill-workbench-rail',
       'maka-streaming-token-fade-in',
       'settingsCardProviders',
       'settingsFeatureStatusHeroActions',
