+++ packages/ui/src/index.ts
@@ -35,6 +35,7 @@ export * from './primitives/menu.js';
 export * from './primitives/group.js';
 export * from './primitives/frame.js';
 export * from './primitives/preview-card.js';
+export * from './primitives/settings-select.js';
 export * from './primitives/input-group.js';
 export * from './primitives/pagination.js';
 export * from './primitives/sidebar.js';
